refactor: move BuildToolDelegatingCommand to cli-common.
