+++ devtools/cli-common/src/main/java/io/quarkus/cli/common/registry/ToggleRegistryClientMixin.java
@@ -1,4 +1,4 @@
-package io.quarkus.cli.registry;
+package io.quarkus.cli.common.registry;
 
 import picocli.CommandLine;
