+++ devtools/cli/src/main/java/io/quarkus/cli/BuildReport.java
@@ -11,7 +11,7 @@ import java.util.Set;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import io.quarkus.cli.build.BuildSystemRunner;
+import io.quarkus.cli.common.build.BuildSystemRunner;
 import io.quarkus.qute.CheckedTemplate;
 import io.quarkus.qute.TemplateInstance;
