+++ devtools/cli/src/main/java/io/quarkus/cli/Deploy.java
@@ -2,6 +2,7 @@ package io.quarkus.cli;
 
 import java.util.Map;
 
+import io.quarkus.cli.common.BuildToolDelegatingCommand;
 import io.quarkus.cli.deploy.Kind;
 import io.quarkus.cli.deploy.Knative;
 import io.quarkus.cli.deploy.Kubernetes;
