+++ devtools/cli/src/main/java/io/quarkus/cli/Dev.java
@@ -7,10 +7,9 @@ import java.util.TreeMap;
 import java.util.concurrent.Callable;
 import java.util.function.Supplier;
 
-import io.quarkus.cli.build.BaseBuildCommand;
-import io.quarkus.cli.build.BuildSystemRunner;
 import io.quarkus.cli.common.DebugOptions;
 import io.quarkus.cli.common.DevOptions;
+import io.quarkus.cli.common.build.BuildSystemRunner;
 import io.quarkus.devtools.project.BuildTool;
 import picocli.CommandLine;
 import picocli.CommandLine.Parameters;
