+++ devtools/cli/src/main/java/io/quarkus/cli/QuarkusCli.java
@@ -23,13 +23,13 @@ import io.quarkus.cli.common.OutputProvider;
 import io.quarkus.cli.common.PropertiesOptions;
 import io.quarkus.cli.common.TargetQuarkusPlatformGroup;
 import io.quarkus.cli.common.VersionHelper;
+import io.quarkus.cli.common.registry.RegistryClientMixin;
 import io.quarkus.cli.plugin.Plugin;
 import io.quarkus.cli.plugin.PluginCommandFactory;
 import io.quarkus.cli.plugin.PluginListItem;
 import io.quarkus.cli.plugin.PluginListTable;
 import io.quarkus.cli.plugin.PluginManager;
 import io.quarkus.cli.plugin.PluginManagerSettings;
-import io.quarkus.cli.registry.RegistryClientMixin;
 import io.quarkus.devtools.project.BuildTool;
 import io.quarkus.devtools.project.QuarkusProject;
 import io.quarkus.devtools.project.QuarkusProjectHelper;
