+++ devtools/cli/src/main/java/io/quarkus/cli/Test.java
@@ -7,13 +7,12 @@ import java.util.TreeMap;
 import java.util.concurrent.Callable;
 import java.util.function.Supplier;
 
-import io.quarkus.cli.build.BaseBuildCommand;
-import io.quarkus.cli.build.BuildSystemRunner;
-import io.quarkus.cli.build.BuildSystemRunner.BuildCommandArgs;
 import io.quarkus.cli.common.BuildOptions;
 import io.quarkus.cli.common.DebugOptions;
 import io.quarkus.cli.common.DevOptions;
 import io.quarkus.cli.common.RunModeOption;
+import io.quarkus.cli.common.build.BuildSystemRunner;
+import io.quarkus.cli.common.build.BuildSystemRunner.BuildCommandArgs;
 import io.quarkus.devtools.project.BuildTool;
 import picocli.CommandLine;
 import picocli.CommandLine.Parameters;
