+++ devtools/cli/src/main/java/io/quarkus/cli/deploy/Kind.java
@@ -2,7 +2,7 @@ package io.quarkus.cli.deploy;
 
 import java.util.Optional;
 
-import io.quarkus.cli.BuildToolContext;
+import io.quarkus.cli.common.BuildToolContext;
 import picocli.CommandLine;
 
 @CommandLine.Command(name = "kind", sortOptions = false, showDefaultValues = true, mixinStandardHelpOptions = false, header = "Perform the deploy action on Kind.", description = "%n"
