+++ devtools/cli/src/main/java/io/quarkus/cli/image/BaseImageSubCommand.java
@@ -2,7 +2,7 @@ package io.quarkus.cli.image;
 
 import java.util.Optional;
 
-import io.quarkus.cli.BuildToolDelegatingCommand;
+import io.quarkus.cli.common.BuildToolDelegatingCommand;
 import picocli.CommandLine.ParentCommand;
 
 public class BaseImageSubCommand extends BaseImageCommand {
