+++ devtools/cli/src/main/java/io/quarkus/cli/image/Docker.java
@@ -3,7 +3,7 @@ package io.quarkus.cli.image;
 import java.util.Map;
 import java.util.Optional;
 
-import io.quarkus.cli.BuildToolContext;
+import io.quarkus.cli.common.BuildToolContext;
 import picocli.CommandLine;
 
 @CommandLine.Command(name = "docker", sortOptions = false, showDefaultValues = true, mixinStandardHelpOptions = false, header = "Build a container image using Docker.", description = "%n"
